[phpdocs] adds phpdocs and descriptions

+++ src/Exception/AmountOutOfRangeException.php
@@ -4,9 +4,19 @@ declare(strict_types=1);
 
 namespace Lendable\Interview\Interpolation\Exception;
 
+/**
+ * Class AmountOutOfRangeException
+ *
+ * @package Lendable\Interview\Interpolation\Exception
+ */
 class AmountOutOfRangeException extends BadRequestException
 {
-    public function __construct(int $loanAmount)
+    /**
+     * AmountOutOfRangeException constructor.
+     *
+     * @param float $loanAmount
+     */
+    public function __construct(float $loanAmount)
     {
         parent::__construct('The loan amount ' . $loanAmount . ' is out of the term range defined');
     }
